Persistence State + Splash Page Re-design
If you look at react dev tools, the state is actually being passed into the container component CreateUserPage. However, the form data is not persisting when using the previous button. This was fixed by transforming the switchcard from a component to a function that still returns JSX but inside of the CreateUserCard for now.

The props currentForm is what calls the form data being passed into whichever step the page is on for the current page. I'm too tired to explain more but it works now.

-also added the object merge for all the form data
-also added the ability to hit the post request to server but has to be hooked up to the server's data.