Login validated using formik
Login validated using formik, classes for input match the bootstrap style.

Logic for it to hit server is not included though. It does use a function called onSubmitUserLogin that can be used to hit the server then retrieved the necessary information.

As, to how it picks up and sends that information would  either have to use a cookie to store current session as being validated then allow the page to be viewied or to always have the pages be allowed to be viewable. And thus everytime a person goes to the page it will simply do a get request from the server for the information.

The user would need a unique id though. Using just the family name, firstname, and lastname will probably not be good when retrieving data.